ge: graphics: pipeline: move pipeline config to a separate file
