+++ include/genesis/assets/pipeline_resource.h
@@ -34,10 +34,12 @@
 
 #include <genesis/assets/resource_base.h>
 #include <genesis/core/memory.h>
-#include <genesis/graphics/pipeline.h>
+#include <genesis/graphics/pipeline_config.h>
 
 namespace GE {
+class Pipeline;
 class Renderer;
+class Shader;
 } // namespace GE
 
 namespace GE::Assets {
