+++ src/genesis/graphics/vulkan/pipeline_config.cpp
@@ -30,7 +30,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "blending.h"
+#include "pipeline_config.h"
 
 #include "genesis/core/asserts.h"
