+++ src/genesis/graphics/vulkan/pipeline_config.h
@@ -32,11 +32,30 @@
 
 #pragma once
 
-#include <genesis/graphics/blending.h>
+#include <genesis/graphics/pipeline_config.h>
 
 #include <vulkan/vulkan.h>
 
 namespace GE::Vulkan {
+
+class DescriptorPool;
+
+struct pipeline_config_t: GE::pipeline_config_t {
+    VkPipelineCache pipeline_cache{VK_NULL_HANDLE};
+    std::vector<VkFormat> color_formats{VK_FORMAT_UNDEFINED};
+    VkFormat depth_format{VK_FORMAT_UNDEFINED};
+    VkFrontFace front_face{VK_FRONT_FACE_COUNTER_CLOCKWISE};
+    VkSampleCountFlagBits msaa_samples{VK_SAMPLE_COUNT_1_BIT};
+    Shared<DescriptorPool> descriptor_pool{};
+
+    VkPipelineViewportStateCreateInfo viewport_state{};
+    VkPipelineInputAssemblyStateCreateInfo input_assembly_state{};
+    VkPipelineRasterizationStateCreateInfo rasterization_state{};
+    VkPipelineMultisampleStateCreateInfo multisample_state{};
+    VkPipelineDepthStencilStateCreateInfo depth_stencil_state{};
+    std::vector<VkDynamicState> dynamic_state_list{};
+    VkPipelineDynamicStateCreateInfo dynamic_state{};
+};
 
 GE_API VkBlendFactor toVkBlendFactor(BlendFactor factor);
 GE_API VkBlendOp toVkBlendOp(BlendOp op);
