+++ src/genesis/scene/entity_picker.cpp
@@ -39,6 +39,8 @@
 
 #include "genesis/assets/resource_id.h"
 #include "genesis/graphics/framebuffer.h"
+#include "genesis/graphics/pipeline_config.h"
+#include "genesis/graphics/shader.h"
 #include "genesis/graphics/staging_buffer.h"
 
 namespace GE::Scene {
