+++ src/genesis/scene/renderer/wb_oit_renderer.cpp
@@ -38,7 +38,9 @@
 #include "genesis/assets/registry.h"
 #include "genesis/graphics/framebuffer.h"
 #include "genesis/graphics/graphics.h"
+#include "genesis/graphics/pipeline_config.h"
 #include "genesis/graphics/renderer.h"
+#include "genesis/graphics/shader.h"
 
 namespace GE::Scene {
 namespace {
